+++ internal/router/inbound_order.go
@@ -2,7 +2,7 @@ package router
 
 import (
 	"github.com/go-chi/chi/v5"
-	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler"
+	handler "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/inbound_order"
 )
 
 // Monta las rutas relacionadas con la entidad Inbound_Orders en el router API principal.
